configを分離
