+++ internal/database/database.go
@@ -3,51 +3,19 @@ package database
 import (
 	"fmt"
 	"log"
-	"os"
 
+	"github.com/fun-dotto/announcement-api/internal/config"
 	"gorm.io/driver/postgres"
 	"gorm.io/gorm"
 )
 
-// データベース設定
-type Config struct {
-	// 接続設定
-	Host     string
-	Port     string
-	User     string
-	Password string
-	DBName   string
-	SSLMode  string
-}
-
-// 環境変数から読み込んだデフォルトのデータベース設定を取得
-func DefaultConfig() *Config {
-	return &Config{
-		Host:     getEnv("DB_HOST", "localhost"),
-		Port:     getEnv("DB_PORT", "5432"),
-		User:     getEnv("DB_USER", "postgres"),
-		Password: getEnv("DB_PASSWORD", "postgres"),
-		DBName:   getEnv("DB_NAME", "postgres"),
-		SSLMode:  getEnv("DB_SSLMODE", "disable"),
-	}
-}
-
-// PostgreSQLの接続文字列を取得
-func (c *Config) DSN() string {
-	return fmt.Sprintf(
+func Connect(cfg config.DatabaseConfig) (*gorm.DB, error) {
+	dsn := fmt.Sprintf(
 		"host=%s port=%s user=%s password=%s dbname=%s sslmode=%s",
-		c.Host, c.Port, c.User, c.Password, c.DBName, c.SSLMode,
+		cfg.Host, cfg.Port, cfg.User, cfg.Password, cfg.DBName, cfg.SSLMode,
 	)
-}
 
-// デフォルト設定を使用してデータベース接続を確立
-func Connect() (*gorm.DB, error) {
-	return ConnectWithConfig(DefaultConfig())
-}
-
-// 指定された設定を使用してデータベース接続を確立
-func ConnectWithConfig(cfg *Config) (*gorm.DB, error) {
-	db, err := gorm.Open(postgres.Open(cfg.DSN()), &gorm.Config{})
+	db, err := gorm.Open(postgres.Open(dsn), &gorm.Config{})
 	if err != nil {
 		return nil, fmt.Errorf("failed to connect to database: %w", err)
 	}
@@ -72,12 +40,4 @@ func Close(db *gorm.DB) error {
 	}
 
 	return nil
-}
-
-// 環境変数を取得するか、デフォルト値を返す
-func getEnv(key, defaultValue string) string {
-	if value := os.Getenv(key); value != "" {
-		return value
-	}
-	return defaultValue
 }
